requirements.txt
Flask~=2.2.5
gunicorn==20.1.0
PyJWT==2.8.0
flask
supabase~=2.18.1
python-dotenv~=1.0.0

fastapi~=0.116.2
pydantic~=2.11.9